Reorganize a few extension method
