+++ Mapsui.Rendering.Skia/SkiaWidgets/IconButtonWidgetRenderer.cs
@@ -1,5 +1,4 @@
 ﻿using Mapsui.Rendering.Skia.Extensions;
-using Mapsui.Rendering.Skia.Images;
 using Mapsui.Widgets;
 using Mapsui.Widgets.ButtonWidgets;
 using SkiaSharp;
